Issue with semi shallow copy solved
Previously, to import I added an attribute name_ctg to the toscaType.
This forced me to create semi shallow copies of the object when
duplicating it with differnet names. But this can lead to a problem
if we modify the element in the future since it wouldn't modify the
other copies unless it's a deep modification.
Now the full names, short names, etc are all linked to the same object.
So there won't be this kind of issue anymore.