Substitution_mappings and tests
Architecture of the function to substitute the abstract nodes
Substitution of abstract node with the directive "substitute" or "substitutable"
(using the substitution_mapping mechanic basically) implemented.
Tests of substitutions of abstract nodes using "substitute" done

Problems: Type resolution not done yet therefore, we manipulate strings instead of the real objects.
(TO DO once the type resolution is done)